+++ spec/griddler/email_spec.rb
@@ -9,7 +9,7 @@ describe Griddler::Email, 'body formatting' do
       <p>Hello.</p><span>Reply ABOVE THIS LINE</span><p>original message</p>
     EOF
 
-    body_from_email(html: body).should eq 'Hello.'
+    expect(body_from_email(html: body)).to eq 'Hello.'
   end
 
   it 'uses the html field and sanitizes it when text param is empty' do
@@ -17,31 +17,31 @@ describe Griddler::Email, 'body formatting' do
       <p>Hello.</p><span>Reply ABOVE THIS LINE</span><p>original message</p>
     EOF
 
-    body_from_email(html: body, text: '').should eq 'Hello.'
+    expect(body_from_email(html: body, text: '')).to eq 'Hello.'
   end
 
   it 'handles invalid utf-8 bytes in html' do
-    body_from_email(html: "Hell\xC0.").should eq 'HellÀ.'
+    expect(body_from_email(html: "Hell\xC0.")).to eq 'HellÀ.'
   end
 
   it 'handles invalid utf-8 bytes in text' do
-    body_from_email(text: "Hell\xF6.").should eq 'Hellö.'
+    expect(body_from_email(text: "Hell\xF6.")).to eq 'Hellö.'
   end
 
   it 'handles valid utf-8 bytes in html' do
-    body_from_email(html: "Hell\xF1.").should eq 'Hellñ.'
+    expect(body_from_email(html: "Hell\xF1.")).to eq 'Hellñ.'
   end
 
   it 'handles valid utf-8 bytes in text' do
-    body_from_email(text: "Hell\xF2.").should eq 'Hellò.'
+    expect(body_from_email(text: "Hell\xF2.")).to eq 'Hellò.'
   end
 
   it 'handles valid utf-8 char in html' do
-    body_from_email(html: 'Hellö.').should eq 'Hellö.'
+    expect(body_from_email(html: 'Hellö.')).to eq 'Hellö.'
   end
 
   it 'handles valid utf-8 char in text' do
-    body_from_email(text: 'Hellö.').should eq 'Hellö.'
+    expect(body_from_email(text: 'Hellö.')).to eq 'Hellö.'
   end
 
   it 'does not remove invalid utf-8 bytes if charset is set' do
@@ -53,7 +53,7 @@ describe Griddler::Email, 'body formatting' do
       text: 'iso-8859-1'
     }
 
-    body_from_email({ text: 'Helló.' }, charsets).should eq 'Helló.'
+    expect(body_from_email({ text: 'Helló.' }, charsets)).to eq 'Helló.'
   end
 
   it 'handles everything on one line' do
@@ -61,7 +61,7 @@ describe Griddler::Email, 'body formatting' do
       Hello. On 01/12/13, Tristan <email@example.com> wrote: Reply ABOVE THIS LINE or visit your website to respond.
     EOF
 
-    body_from_email(text: body).should eq 'Hello.'
+    expect(body_from_email(text: body)).to eq 'Hello.'
   end
 
   it 'handles "On [date] [soandso] wrote:" format' do
@@ -76,7 +76,7 @@ describe Griddler::Email, 'body formatting' do
       > Thanks, Tristan
     EOF
 
-    body_from_email(text: body).should eq 'Hello.'
+    expect(body_from_email(text: body)).to eq 'Hello.'
   end
 
   it 'handles "On [date] [soandso] <email@example.com> wrote:" format' do
@@ -91,7 +91,7 @@ describe Griddler::Email, 'body formatting' do
       > Thanks, Tristan
     EOF
 
-    body_from_email(text: body).should eq 'Hello.'
+    expect(body_from_email(text: body)).to eq 'Hello.'
   end
 
   it 'handles "On [date] [soandso]\n<email@example.com> wrote:" format' do
@@ -106,7 +106,7 @@ describe Griddler::Email, 'body formatting' do
       > Thanks, Tristan
     EOF
 
-    body_from_email(text: body).should eq 'Hello.'
+    expect(body_from_email(text: body)).to eq 'Hello.'
   end
 
   it 'handles "> On [date] [soandso] <email@example.com> wrote:" format' do
@@ -122,7 +122,7 @@ describe Griddler::Email, 'body formatting' do
       >
     EOF
 
-    body_from_email(text: body).should eq 'Hello.'
+    expect(body_from_email(text: body)).to eq 'Hello.'
   end
 
   it 'handles "From: email@email.com" format' do
@@ -136,7 +136,7 @@ describe Griddler::Email, 'body formatting' do
       Check out this report!
     EOF
 
-    body_from_email(text: body).should eq 'Hello.'
+    expect(body_from_email(text: body)).to eq 'Hello.'
   end
 
   it 'handles "-----Original Message-----" format' do
@@ -151,7 +151,7 @@ describe Griddler::Email, 'body formatting' do
       Check out this report!
     EOF
 
-    body_from_email(text: body).should eq 'Hello.'
+    expect(body_from_email(text: body)).to eq 'Hello.'
   end
 
   it 'handles "-----Original Message-----" format without a preceding body' do
@@ -164,7 +164,7 @@ describe Griddler::Email, 'body formatting' do
       Check out this report!
     EOF
 
-    body_from_email(text: body).should eq ''
+    expect(body_from_email(text: body)).to eq ''
   end
 
   it 'handles "-----Original message-----" case insensitively' do
@@ -179,7 +179,7 @@ describe Griddler::Email, 'body formatting' do
       Check out this report!
     EOF
 
-    body_from_email(text: body).should eq 'Hello.'
+    expect(body_from_email(text: body)).to eq 'Hello.'
   end
 
   it 'handles "-----Original message-----" case insensitively without a preceding body' do
@@ -192,7 +192,7 @@ describe Griddler::Email, 'body formatting' do
       Check out this report!
     EOF
 
-    body_from_email(text: body).should eq ''
+    expect(body_from_email(text: body)).to eq ''
   end
 
   it 'handles "[date] [soandso] <email@example>" format' do
@@ -205,7 +205,7 @@ describe Griddler::Email, 'body formatting' do
       > Thanks, Tristan
     EOF
 
-    body_from_email(text: body).should eq ''
+    expect(body_from_email(text: body)).to eq ''
   end
 
   it 'handles "Reply ABOVE THIS LINE" format' do
@@ -217,7 +217,7 @@ describe Griddler::Email, 'body formatting' do
       Hey!
     EOF
 
-    body_from_email(text: body).should eq 'Hello.'
+    expect(body_from_email(text: body)).to eq 'Hello.'
   end
 
   it 'removes > in "> Reply ABOVE THIS LINE" ' do
@@ -227,7 +227,7 @@ describe Griddler::Email, 'body formatting' do
       > Reply ABOVE THIS LINE
     EOF
 
-    body_from_email(text: body).should eq 'Hello.'
+    expect(body_from_email(text: body)).to eq 'Hello.'
   end
 
   it 'removes any non-content things above Reply ABOVE THIS LINE' do
@@ -241,7 +241,7 @@ describe Griddler::Email, 'body formatting' do
       Hey!
     EOF
 
-    body_from_email(text: body).should eq 'Hello.'
+    expect(body_from_email(text: body)).to eq 'Hello.'
   end
 
   it 'removes any iphone things above Reply ABOVE THIS LINE' do
@@ -255,7 +255,7 @@ describe Griddler::Email, 'body formatting' do
       Hey!
     EOF
 
-    body_from_email(text: body).should eq 'Hello.'
+    expect(body_from_email(text: body)).to eq 'Hello.'
   end
 
   it 'should remove any signature above Reply ABOVE THIS LINE' do
@@ -272,7 +272,7 @@ describe Griddler::Email, 'body formatting' do
       > Hey!
     EOF
 
-    body_from_email(text: body).should eq 'Hello.'
+    expect(body_from_email(text: body)).to eq 'Hello.'
   end
 
   it 'should remove any signature without space above Reply ABOVE THIS LINE' do
@@ -289,7 +289,7 @@ describe Griddler::Email, 'body formatting' do
       > Hey!
     EOF
 
-    body_from_email(text: body).should eq 'Hello.'
+    expect(body_from_email(text: body)).to eq 'Hello.'
   end
 
   it 'properly handles a json charsets' do
@@ -314,27 +314,27 @@ describe Griddler::Email, 'body formatting' do
       text: 'utf-8'
     }
 
-    body_from_email({ text: body }, charsets).should eq 'Hello.'
+    expect(body_from_email({ text: body }, charsets)).to eq 'Hello.'
   end
 
   it 'should preserve empty lines' do
     body = "Hello.\n\nWhat's up?"
 
-    body_from_email(text: body).should eq body
+    expect(body_from_email(text: body)).to eq body
   end
 
   it 'preserves blockquotes' do
     body = "> Hello.\n\n>another line"
 
-    body_from_email(text: body).should eq body
+    expect(body_from_email(text: body)).to eq body
   end
 
   it 'handles empty body values' do
-    body_from_email(text: '').should eq ''
+    expect(body_from_email(text: '')).to eq ''
   end
 
   it 'handles missing body keys' do
-    body_from_email(text: nil).should eq ''
+    expect(body_from_email(text: nil)).to eq ''
   end
 
   def body_from_email(raw_body, charsets = {})
@@ -365,8 +365,8 @@ describe Griddler::Email, 'multipart emails' do
       html: '<b>hello there</b>',
       text: 'hello there'
     )
-    email.raw_html.should eq '<b>hello there</b>'
-    email.raw_text.should eq 'hello there'
+    expect(email.raw_html).to eq '<b>hello there</b>'
+    expect(email.raw_text).to eq 'hello there'
   end
 
   it 'uses text as raw_body if both text and html are present' do
@@ -374,21 +374,21 @@ describe Griddler::Email, 'multipart emails' do
       html: '<b>hello there</b>',
       text: 'hello there'
     )
-    email.raw_body.should eq 'hello there'
+    expect(email.raw_body).to eq 'hello there'
   end
 
   it 'uses text as raw_body' do
     email = email_with_params(
       text: 'hello there'
     )
-    email.raw_body.should eq 'hello there'
+    expect(email.raw_body).to eq 'hello there'
   end
 
   it 'uses html as raw_body if text is not present' do
     email = email_with_params(
       html: '<b>hello there</b>'
     )
-    email.raw_body.should eq '<b>hello there</b>'
+    expect(email.raw_body).to eq '<b>hello there</b>'
   end
 
   it 'uses html as raw_body if text is empty' do
@@ -396,7 +396,7 @@ describe Griddler::Email, 'multipart emails' do
       html: '<b>hello there</b>',
       text: ''
     )
-    email.raw_body.should eq '<b>hello there</b>'
+    expect(email.raw_body).to eq '<b>hello there</b>'
   end
 
   def email_with_params(params)
@@ -416,17 +416,17 @@ describe Griddler::Email, 'extracting email headers' do
     header = "#{header_name}: #{header_value}"
 
     headers = header_from_email(header)
-    headers[header_name].should eq header_value
+    expect(headers[header_name]).to eq header_value
   end
 
   it 'handles no matched headers' do
     headers = header_from_email('')
-    headers.should eq({})
+    expect(headers).to eq({})
   end
 
   it 'handles nil headers' do
     headers = header_from_email(nil)
-    headers.should eq({})
+    expect(headers).to eq({})
   end
 
   def header_from_email(header)
@@ -459,7 +459,7 @@ describe Griddler::Email, 'extracting email addresses' do
       to: [@full_address],
       from: @full_address,
     )
-    email.to.should eq [@address_components.merge(name: 'Bob')]
+    expect(email.to).to eq [@address_components.merge(name: 'Bob')]
   end
 
   it 'handles normal e-mail address' do
@@ -472,16 +472,16 @@ describe Griddler::Email, 'extracting email addresses' do
       full: @address_components[:email],
       name: nil,
     )
-    email.to.should eq [expected]
-    email.from.should eq @address_components
+    expect(email.to).to eq [expected]
+    expect(email.from).to eq @address_components
   end
 
   it 'handles new lines' do
     email = Griddler::Email.new(text: 'hi', to: ["#{@full_address}\n"],
       from: "#{@full_address}\n")
     expected = @address_components.merge(full: "#{@full_address}\n")
-    email.to.should eq [expected]
-    email.from.should eq expected
+    expect(email.to).to eq [expected]
+    expect(email.from).to eq expected
   end
 
   it 'handles angle brackets around address' do
@@ -493,8 +493,8 @@ describe Griddler::Email, 'extracting email addresses' do
     expected = @address_components.merge(
       full: "<#{@address_components[:email]}>",
       name: nil)
-    email.to.should eq [expected]
-    email.from.should eq expected
+    expect(email.to).to eq [expected]
+    expect(email.from).to eq expected
   end
 
   it 'handles name and angle brackets around address' do
@@ -503,8 +503,8 @@ describe Griddler::Email, 'extracting email addresses' do
       to: [@full_address],
       from: @full_address
     )
-    email.to.should eq [@address_components]
-    email.from.should eq @address_components
+    expect(email.to).to eq [@address_components]
+    expect(email.from).to eq @address_components
   end
 
   it 'handles multiple e-mails, with priority to the bracketed' do
@@ -518,8 +518,8 @@ describe Griddler::Email, 'extracting email addresses' do
       name: 'fake@example.com'
     )
 
-    email.to.should eq [expected]
-    email.from.should eq expected
+    expect(email.to).to eq [expected]
+    expect(email.from).to eq expected
   end
 end
 
@@ -531,7 +531,7 @@ describe Griddler::Email, 'extracting email addresses from CC field' do
 
   it 'uses the cc from the adapter' do
     email = Griddler::Email.new(to: [@address], from: @address, cc: [@cc], headers: @headers)
-    email.cc.should eq [{
+    expect(email.cc).to eq [{
       token: 'charles+123',
       host: 'example.com',
       email: 'charles+123@example.com',
@@ -542,7 +542,7 @@ describe Griddler::Email, 'extracting email addresses from CC field' do
 
   it 'returns an empty array when no CC address is added' do
     email = Griddler::Email.new(to: [@address], from: @address)
-    email.cc.should be_empty
+    expect(email.cc).to be_empty
   end
 end
 
@@ -568,10 +568,10 @@ describe Griddler::Email, 'with custom configuration' do
 
   describe 'accepts and works with a string reply delimiter' do
     it 'does not split on Reply ABOVE THIS LINE' do
-      Griddler.configuration.stub(reply_delimiter: 'Stuff and things')
+      allow(Griddler.configuration).to receive_messages(reply_delimiter: 'Stuff and things')
       email = Griddler::Email.new(params)
 
-      email.body.should eq params[:text]
+      expect(email.body).to eq params[:text]
     end
 
     it 'splits at custom delimeter' do
@@ -583,15 +583,15 @@ describe Griddler::Email, 'with custom configuration' do
         wut
       EOS
 
-      Griddler.configuration.stub(reply_delimiter: '-- reply above --')
+      allow(Griddler.configuration).to receive_messages(reply_delimiter: '-- reply above --')
       email = Griddler::Email.new(params)
-      email.body.should eq 'trolololo'
+      expect(email.body).to eq 'trolololo'
     end
   end
 
   describe 'accepts and works with an array of reply delimiters' do
     before do
-      Griddler.configuration.stub(reply_delimiter: ['-- old reply above --', '-- new reply above --'])
+      allow(Griddler.configuration).to receive_messages(reply_delimiter: ['-- old reply above --', '-- new reply above --'])
     end
 
     it 'splits with old delimiter' do
@@ -604,7 +604,7 @@ describe Griddler::Email, 'with custom configuration' do
       EOS
 
       email = Griddler::Email.new(params)
-      email.body.should eq 'Hey, split me with the old one!'
+      expect(email.body).to eq 'Hey, split me with the old one!'
     end
 
     it 'splits with the new delimiter' do
@@ -617,7 +617,7 @@ describe Griddler::Email, 'with custom configuration' do
       EOS
 
       email = Griddler::Email.new(params)
-      email.body.should eq 'Hey, split me with the new one!'
+      expect(email.body).to eq 'Hey, split me with the new one!'
     end
   end
 
@@ -627,7 +627,7 @@ describe Griddler::Email, 'with custom configuration' do
 This is the real text\r\n\r\n\r\nOn Fri, Mar 21, 2014 at 3:11 PM, Someone <\r\nsomeone@example.com> wrote:\r\n\r\n>  -- REPLY ABOVE THIS LINE --\r\n>\r\n> The Old message!\r\n>\r\n> Another line! *\r\n>\n
       EOS
       email = Griddler::Email.new(params)
-      email.body.should eq 'This is the real text'
+      expect(email.body).to eq 'This is the real text'
     end
   end
 
@@ -638,7 +638,7 @@ This is the real text\r\n\r\n\r\nOn Fri, Mar 21, 2014 at 3:11 PM, Someone <\r\ns
 
       email = Griddler::Email.new(params)
 
-      email.to.map { |to| to[:full] }.should eq recipients
+      expect(email.to.map { |to| to[:full] }).to eq recipients
     end
   end
 end